requirements.txt
# Core dependencies
google-genai>=1.0.0
firebase-admin>=6.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image processing
Pillow>=10.0.0

# Mermaid diagram rendering
mermaid-py>=0.5.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
aiofiles>=23.0.0
jinja2>=3.1.0
